rno4: wire psp water effects TU via shared header extracted from rno3
Extract the seven water effect entities from the rno3 us TU into a
shared water_effects_rev.h and rebuild the rno4 psp stub as a
composition file over it, keeping the rno3 us TU byte-identical.
