+++ src/st/rno4_psp/unk_12B08.c
@@ -1,16 +1,4 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 #include "../rno4/rno4.h"
 
-INCLUDE_ASM("st/rno4_psp/nonmatchings/rno4_psp/unk_12B08", func_pspeu_0924B480_from_rno3);
-
-INCLUDE_ASM("st/rno4_psp/nonmatchings/rno4_psp/unk_12B08", EntityAlucardWaterEffect);
-
-INCLUDE_ASM("st/rno4_psp/nonmatchings/rno4_psp/unk_12B08", EntitySplashWater);
-
-INCLUDE_ASM("st/rno4_psp/nonmatchings/rno4_psp/unk_12B08", EntitySurfacingWater);
-
-INCLUDE_ASM("st/rno4_psp/nonmatchings/rno4_psp/unk_12B08", EntitySideWaterSplash);
-
-INCLUDE_ASM("st/rno4_psp/nonmatchings/rno4_psp/unk_12B08", EntitySmallWaterDrop);
-
-INCLUDE_ASM("st/rno4_psp/nonmatchings/rno4_psp/unk_12B08", EntityWaterDrop);
+#include "../water_effects_rev.h"
